Limit error skipping to access errors only.
Skip only access errors, but not any other error type; for consistency, skip
access errors under all platforms.

Silently skipping file access errors in `get_tree()` is safe in the sense that
it does not surprise the user, as the output of `boar status` and that of any
other `get_tree()` callers is consistent.